refactor: break state.ts type-import cycles via types.ts
state.ts imported McpStatusEventBus from mcp-status.ts and UiServerHandle
from ui-server.ts while both modules import McpExtensionState back —
type-level cycles with no runtime effect today, but any future value
import on either edge turns them into real undefined-binding cycles. Both
interfaces now live in types.ts (the leaf module that already hosts the
state-adjacent types they reference), so the state type and the
status/ui-server modules can evolve independently. Verified with
gitnexus check_import_cycles: 0 cycles (a fresh index also surfaced the
ui-server edge, which the original stale index had missed).
